requirements.txt
pyspark==3.4.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
pytest-mock==3.11.1
pandas==2.0.3

